没有实力提升的感觉

+++ Algorithm/luogu/dp 1005.cpp
@@ -9,8 +9,8 @@ LL two[200];
 
 
 
-
-int dfs(int b,int e,int t)
+// int dfs(int b,int e,int t)//一开始还写成这个屌样
+LL dfs(int b,int e,int t)
 {
     LL one =  1;
     if(dp[b][e]!=-1)
